examples: Add multi material example
This example demonstrates issues with a combination of flat variables
and sampler arrays.

This seems to happen within `glslopt`.
Without the `flat` specifier, shaders are generated with a version line of
`120` and with the `flat` the versions line is `130`.

The output of `glslopt_optimize` has `texture2DArray` for 120 and
`textureArray` for 130.